pyproject.toml
[project]
name = "hopper-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "msal>=1.34.0",
    "pocketbase>=0.15.0",
    "pydantic[email]>=2.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
